Refactor device blueprint parsing into modular stack
